+++ src/components/crewInput.js
@@ -1,0 +1,1 @@
+export default class CrewInput {}
